requirements.txt
glob2
Flask-Cors
mysql-connector-python
absl-py==0.13.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
Flask==2.0.1
google-auth==1.34.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.1
oauthlib==3.1.1
opencv-python==4.5.3.56
pandas==1.3.1
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
seaborn==0.11.1
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.7.0+cpu
tqdm==4.62.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
